speed up when compare equal
