+++ src/simulation/include/Simulation.hpp
@@ -7,6 +7,7 @@
 #include "ExpiringState.hpp"
 #include "BaseModel.hpp"
 #include "Heap.hpp"
+// #include "util.hpp"
 
 #include <omp.h>
