--- src/simulation/include/util.hpp
@@ -1,10 +1,0 @@
-#ifndef UTIL_HPP
-#define UTIL_HPP
-
-#include <vector>
-#include "Type.hpp"
-
-template<typename T>
-std::vector<uint> smallest(std::vector<T>& vec, uint k);
-
-#endif
